+++ src/components/containers/CauseForm/IconSelector/IconSelector.spec.js
